Rename: 약관 상세 스크린 폴더명 변경
