+++ lib/screens/terms/terms_detail/terms_detail_screen.dart
